Se terminaron todos los test de dominio y mvc

+++ presupuesto/aplicacion/src/main/java/com/ceiba/gasto/comando/ComandoGasto.java
@@ -15,6 +15,6 @@ public class ComandoGasto {
 
     private Long id;
     private String identificacionUsuario;
-    private int valorGasto;
+    private Long valorGasto;
     private LocalDateTime fechaGasto;
 }
